Reduced Sample Space
Brought number of boards from 390k down to 19k, by disregarding order of
tiles. We only care about the tiling.
This helps the AI learn faster, because it now evaluates based on the
exact board state, rather than needing to have the same state + order of
play.

As a result, needed to add a BoardStorage class. This class is based
with a hashmap, and then if the board hasn't been created it creates it.

Next step is Bot play styles and a way to save/load what the bot has
learned.